refactor(api): implement better api version and use it for all routes
