+++ services/api/app/Http/Controllers/V1/FeatureController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\V1;
 
+use App\Http\Controllers\Controller;
 use App\Http\Resources\FeatureResource;
 use Laravel\Pennant\Feature;
